+++ packages/web-inspector/vitest.setup.ts
@@ -18,39 +18,10 @@
 
 import { beforeEach } from "vitest";
 
-// ---------------------------------------------------------------------------
-// Telemetry egress guard (OSS-565)
-//
-// The inspector's telemetry is browser-side: it keys off `telemetryDisabled`
-// from the runtime's /info handshake, so COPILOTKIT_TELEMETRY_DISABLED in a CI
-// job cannot reach it. Any test that drives a banner / threads / open code
-// path without stubbing fetch therefore POSTs a real `oss.inspector.*` event
-// to the live sink from CI — pollution indistinguishable from user activity.
-//
-// This wrapper swallows requests to the sink and delegates everything else, so
-// tests that stub or spy on fetch themselves are unaffected.
-// ---------------------------------------------------------------------------
+import { createTelemetryEgressGuard } from "./src/lib/testing/telemetry-egress-guard.js";
 
-const TELEMETRY_SINK_ORIGIN = "https://telemetry.copilotkit.ai";
-
-/** Exported for its own unit test — see `telemetry-egress-guard.spec.ts`. */
-export function createTelemetryEgressGuard(
-  realFetch: typeof fetch,
-): typeof fetch {
-  return (input, init) => {
-    const url =
-      typeof input === "string"
-        ? input
-        : input instanceof URL
-          ? input.href
-          : input.url;
-    if (url?.startsWith(TELEMETRY_SINK_ORIGIN)) {
-      return Promise.resolve(new Response(null, { status: 204 }));
-    }
-    return realFetch(input, init);
-  };
-}
-
+// No test may POST a real `oss.inspector.*` event to the live telemetry sink
+// (OSS-565) — see the helper for why a CI env var can't cover this.
 if (typeof globalThis.fetch === "function") {
   globalThis.fetch = createTelemetryEgressGuard(globalThis.fetch);
 }
